Whisper/non real_time_asr requirements.txt
pyaudio
torch
transformers
soundfile
numpy
pydub
pyannote.audio
huggingface_hub
